Refactored log_utils
