+++ core/logging/__init__.py
@@ -5,11 +5,7 @@ logger.setLevel(logging.DEBUG)
 
 formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
 
-# file_handler = logging.FileHandler(filename='debug.log')
-# file_handler.setFormatter(formatter)
-
 stream_handler = logging.StreamHandler()
 stream_handler.setFormatter(formatter)
 
 logger.addHandler(stream_handler)
-# logger.addHandler(file_handler)
